Update e019-rear.py
For some reason there returned the error "to many vales to unpack" in the function "result", and there was a problem with the right reading in of "small_dataset", which might be caused because I just copyed and pasted the code.
The proposed two changes, though perhaps not pretty, are supposed to be able to fix the problems above.
Thank you for your code~ (Though I didn't find the code for Python 3, which is required for my work. But the differences can be easily dealed with.) 
I think the hard part of this problem is the math.